requirements.txt
youtube-transcript-api
google-api-python-client
bs4
requests
ebbe
bloomfilter-py